+++ test/test_documentation.py
@@ -1,4 +1,4 @@
-from src import compare_answer_no_html
+from answerset import compare_answer_no_html
 
 def test_doc_first_example():
     correct = 'depart, go, leave'
