+++ test/test_lenient_validation.py
@@ -1,4 +1,4 @@
-from src import compare_answer_no_html
+from answerset import compare_answer_no_html
 
 def test_no_missing():
     correct = 'test answer'
